Changes to be committed: 	new file:   DataStructure.h 	modified:   FirstShader.fs 	modified:   FirstShader.vs 	modified:   GLProgram.m 	new file:   ObjLoader.h 	new file:   ObjLoader.m 	modified:   PPNode.h 	modified:   PPNode.m 	new file:   banana.obj 	new file:   complicate.obj 	new file:   data.obj 	new file:   pyramid.obj 	modified:   shaderlib.xcodeproj/project.pbxproj 	modified:   shaderlib.xcodeproj/project.xcworkspace/xcuserdata/realxie.xcuserdatad/UserInterfaceState.xcuserstate 	modified:   shaderlib/ViewController.m

+++ FirstShader.fs
@@ -8,9 +8,9 @@
 
 uniform sampler2D uSampler;
 //
-varying lowp vec2 vTexCoord;
+varying highp vec2 vTexCoord;
 //
-varying lowp vec4 colorVarying;
+varying highp vec4 colorVarying;
 
 void main()
 {
